pyproject.toml
[tool.poetry]
name = "ds-tg-repost"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ds_tg_repost"}]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.1.0"
python-dotenv = "^0.21.0"
python-telegram-bot = "^20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
